+++ src/schemathesis/generation/jsonschema/__init__.py
@@ -1,5 +1,5 @@
-from schemathesis.generation.jsonschema.builder import build
+from schemathesis.generation.jsonschema.builder import EMPTY_STRATEGY, build
 from schemathesis.generation.jsonschema.context import Alphabet, StrategyContext
 from schemathesis.generation.jsonschema.strategy import from_schema
 
-__all__ = ["Alphabet", "StrategyContext", "build", "from_schema"]
+__all__ = ["EMPTY_STRATEGY", "Alphabet", "StrategyContext", "build", "from_schema"]